chore: rename flag converter file to flag_converter to avoid conflicting with ext.commands flag classes
